requirements.txt
autopep8==1.5.6
beautifulsoup4==4.9.3
Bootstrap-Flask==1.5.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Deprecated==1.2.12
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.2
pandas==1.2.3
pycodestyle==2.7.0
PyGithub==1.54.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
rope==0.18.0
six==1.15.0
soupsieve==2.2.1
toml==0.10.2
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
